Refactor code to prevent clashing of filenames in each function app
Also move validation in each function app as first step in process, ahead
of calling postcode index as it may still fail validation and the call out to
azure search for postcode lookup would have been wasted
