+++ InstitutionSearchAPIHttpTrigger/institution/exceptions.py
